+++ packages/atlas/lib/src/marker_icon.dart
@@ -1,4 +1,4 @@
-import 'package:meta/meta.dart';
+import 'package:flutter/foundation.dart';
 
 /// Optional Marker icon.
 class MarkerIcon {
@@ -9,4 +9,15 @@ class MarkerIcon {
   const MarkerIcon({
     @required this.assetName,
   }) : assert(assetName != null);
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other.runtimeType != runtimeType) return false;
+    final MarkerIcon typedOther = other;
+    return assetName == typedOther.assetName;
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode ^ assetName.hashCode;
 }